refactor: moved the LottieDrawable outside the paging adapter and in the fragment
